Power Pin addition
After working with Darthcloud, it was revealed that some PS2 consoles have weak VIN on either the 8V or the 3V3.  To make sure this PCB will work in either situation, I have added the 8V (Pad 3) pad back on to the PCB in addition to the 3V3 Pad (Pad 5).  Only one needs to be used to power the ESP.