Retiring Main PCB
Retiring Original Main PCB design as the unit is thinner that the standard footprint.  This was causing  confusion.  The new main used to be Main Alt Wide version.  It is now the standard as it fits the standard ESP footprint.  The Thinner version has been moved to the retired folder in PCBs