CAD Update
I have completed the preliminary CAD for the Wii EXT & NES/SNES Classic PCBs.  I have ordered them from Shapeways and will merge the branch if no additional changes are needed.

Included in this commit are the following:
Updated CAD and STLs for the AUX housing Top for the NES Classic, SNES Classic, Famicom Classic, Super Famicom Classic, Wii Ext and Oprional Wii

The Aux Bottom that covers all the variants mentioned above.

As Always the CAD is included and the new font for the Wii Variants is included as well.